Refactor, document, and cleanup
