+++ cmd/osv-scanner/internal/cmd/run.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	"github.com/google/osv-scanner/v2/internal/testlogger"
 	"io"
 	"log/slog"
 	"testing"
@@ -22,11 +23,24 @@ var (
 )
 
 func Run(args []string, stdout, stderr io.Writer) int {
-	// Only set the global logHandler if we are not in testing mode
-	if !testing.Testing() {
-		logger := cmdlogger.New(stdout, stderr)
-		slog.SetDefault(slog.New(&logger))
+
+	// --- Setup Logger ---
+	logHandler := cmdlogger.New(stdout, stderr)
+
+	// If in testing mode, set logger via TestLogger
+	// Otherwise, set default global logger
+	if testing.Testing() {
+		handler, ok := slog.Default().Handler().(*testlogger.TestLogger)
+		if !ok {
+			panic("Test failed to initialize default logger with TestLogger")
+		}
+
+		handler.AddInstance(logHandler)
+		defer handler.Delete()
+	} else {
+		slog.SetDefault(slog.New(logHandler))
 	}
+	// ---
 
 	cli.VersionPrinter = func(ctx *cli.Context) {
 		slog.Info("osv-scanner version: " + ctx.App.Version)
@@ -78,7 +92,6 @@ func Run(args []string, stdout, stderr io.Writer) int {
 		slog.Error(fmt.Sprintf("%v", err))
 	}
 
-	logHandler := slog.Default().Handler().(cmdlogger.CmdLogger)
 	// if we've been told to print an error, and not already exited with
 	// a specific error code, then exit with a generic non-zero code
 	if logHandler.HasErrored() {
